history ajax polymer gets url parsed current location
