+++ util/git/struct_test_data.go
@@ -1,0 +1,40 @@
+package gyena
+
+var showCommitText = `commit fc8e7351b01144fa299c23357a1b4b64708684a4
+Author: sosuke-k <snoopies.drum@gmail.com>
+Date:   Mon Jun 29 02:53:10 2015 +0900
+
+    hyena auto git commit
+
+diff --git a/chrome.json b/chrome.json
+index 81bfe46..079f8b2 100644
+--- a/chrome.json
++++ b/chrome.json
+@@ -1,7 +1,6 @@
+ {
+   "0": [
+-    "https://github.com/sosuke-k/hyena",
+-    "http://ejje.weblio.jp/content/hyena"
++    "https://github.com/sosuke-k/hyena/compare/sosuke-k/git?expand=1"
+   ],
+   "1": [
+     "http://aidiary.hatenablog.com/entry/20110514/1305377659",`
+
+var secondCommitLog = `commit 65d8e78f9275df5abebcb94284b1fd49418af517
+Author: sosuke-k <snoopies.drum@gmail.com>
+Date:   Mon Jun 29 02:47:15 2015 +0900
+
+    hyena auto git commit
+
+diff --git a/atom.json b/atom.json
+index 0f7a69f..0acae8f 100644
+--- a/atom.json
++++ b/atom.json
+@@ -1,5 +1,5 @@
+ {
+   "0": [
+-    "/Users/katososuke/go/src/github.com/sosuke-k/hyena"
++    null
+   ]
+ }
+\ No newline at end of file`
